♻️ (donate) move non fn files to utils
